consolidate internal packages
