+++ internal/testdata/bad-scalar.yaml
