+++ internal/testdata/bad-schema.yaml
