+++ internal/testdata/k8s-schema.yaml
