+++ internal/testdata/scalar-compare-output.txt
