+++ internal/testdata/scalar.yaml
