+++ internal/testdata/struct.yaml
